Move StreamLayerClient cache tests to separate file
Move StreamLayerClient cache tests to separate file.

Relates-to: OLPEDGE-720
