+++ tests/functional/CMakeLists.txt
@@ -23,6 +23,7 @@ set(OLP_SDK_FUNCTIONAL_TESTS_SOURCES
     ./olp-cpp-sdk-dataservice-read/ApiTest.cpp
     ./olp-cpp-sdk-dataservice-read/CatalogClientTest.cpp
     ./olp-cpp-sdk-dataservice-write/DataserviceWriteIndexLayerClientTest.cpp
+    ./olp-cpp-sdk-dataservice-write/DataserviceWriteStreamLayerClientCacheTest.cpp
     ./olp-cpp-sdk-dataservice-write/DataserviceWriteStreamLayerClientTest.cpp
     ./olp-cpp-sdk-dataservice-write/DataserviceWriteVersionedLayerClientTest.cpp
     ./olp-cpp-sdk-dataservice-write/DataserviceWriteVolatileLayerClientTest.cpp
