+++ tests/functional/olp-cpp-sdk-dataservice-write/DataserviceWriteStreamLayerClientTest.cpp
@@ -725,536 +725,4 @@ TEST_F(DataserviceWriteStreamLayerClientTest,
   async5.get();
 }
 
-class StreamLayerClientCacheOnlineTest
-    : public DataserviceWriteStreamLayerClientTest {
- protected:
-  static std::shared_ptr<olp::http::Network> s_network;
-
-  static void SetUpTestSuite() {
-    s_network = olp::client::OlpClientSettingsFactory::
-        CreateDefaultNetworkRequestHandler();
-  }
-
-  virtual std::shared_ptr<StreamLayerClient> CreateStreamLayerClient()
-      override {
-    auto network = s_network;
-
-    olp::authentication::Settings authentication_settings;
-    authentication_settings.token_endpoint_url =
-        CustomParameters::getArgument(kEndpoint);
-    authentication_settings.network_request_handler = network;
-
-    olp::authentication::TokenProviderDefault provider(
-        CustomParameters::getArgument(kAppid),
-        CustomParameters::getArgument(kSecret), authentication_settings);
-
-    olp::client::AuthenticationSettings auth_client_settings;
-    auth_client_settings.provider = provider;
-
-    olp::client::OlpClientSettings settings;
-    settings.authentication_settings = auth_client_settings;
-    settings.network_request_handler = network;
-    settings.task_scheduler =
-        olp::client::OlpClientSettingsFactory::CreateDefaultTaskScheduler(1u);
-
-    disk_cache_ = std::make_shared<olp::cache::DefaultCache>();
-    EXPECT_EQ(disk_cache_->Open(),
-              olp::cache::DefaultCache::StorageOpenResult::Success);
-
-    return std::make_shared<StreamLayerClient>(
-        olp::client::HRN{GetTestCatalog()}, settings, disk_cache_,
-        flush_settings_);
-  }
-
-  virtual void TearDown() override {
-    DataserviceWriteStreamLayerClientTest::TearDown();
-    if (disk_cache_) {
-      disk_cache_->Close();
-    }
-  }
-
-  std::shared_ptr<olp::cache::DefaultCache> disk_cache_;
-  FlushSettings flush_settings_;
-};
-
-// Static network instance is necessary as it needs to outlive any created
-// clients. This is a known limitation as triggered send requests capture the
-// network instance inside the callbacks.
-std::shared_ptr<olp::http::Network> StreamLayerClientCacheOnlineTest::s_network;
-
-TEST_F(StreamLayerClientCacheOnlineTest, Queue) {
-  auto error = client_->Queue(
-      PublishDataRequest().WithData(data_).WithLayerId(GetTestLayer()));
-
-  ASSERT_FALSE(error) << error.get();
-}
-
-TEST_F(StreamLayerClientCacheOnlineTest, QueueNullData) {
-  auto error = client_->Queue(
-      PublishDataRequest().WithData(nullptr).WithLayerId(GetTestLayer()));
-
-  ASSERT_TRUE(error);
-}
-
-TEST_F(StreamLayerClientCacheOnlineTest, QueueExtraRequestParams) {
-  const auto uuid = GenerateRandomUUID();
-
-  auto error = client_->Queue(PublishDataRequest()
-                                  .WithData(data_)
-                                  .WithLayerId(GetTestLayer())
-                                  .WithTraceId(uuid)
-                                  .WithBillingTag(kBillingTag));
-
-  ASSERT_FALSE(error) << error.get();
-}
-
-#ifdef DATASERVICE_WRITE_HAS_OPENSSL
-TEST_F(StreamLayerClientCacheOnlineTest, QueueWithChecksum) {
-  const std::string data_string(data_->begin(), data_->end());
-  auto checksum = sha256(data_string);
-
-  auto error = client_->Queue(PublishDataRequest()
-                                  .WithData(data_)
-                                  .WithLayerId(GetTestLayer())
-                                  .WithChecksum(checksum));
-
-  ASSERT_FALSE(error) << error.get();
-}
-#endif
-
-TEST_F(StreamLayerClientCacheOnlineTest, FlushDataSingle) {
-  auto error = client_->Queue(
-      PublishDataRequest().WithData(data_).WithLayerId(GetTestLayer()));
-
-  ASSERT_FALSE(error) << error.get();
-
-  auto response = client_->Flush().GetFuture().get();
-
-  ASSERT_FALSE(response.empty());
-  ASSERT_NO_FATAL_FAILURE(PublishDataSuccessAssertions(response[0]));
-}
-
-TEST_F(StreamLayerClientCacheOnlineTest, FlushDataMultiple) {
-  ASSERT_NO_FATAL_FAILURE(QueueMultipleEvents(5));
-
-  auto response = client_->Flush().GetFuture().get();
-
-  ASSERT_EQ(5, response.size());
-  for (auto& single_response : response) {
-    ASSERT_NO_FATAL_FAILURE(PublishDataSuccessAssertions(single_response));
-  }
-}
-
-TEST_F(StreamLayerClientCacheOnlineTest, FlushDataSingleAsync) {
-  auto error = client_->Queue(
-      PublishDataRequest().WithData(data_).WithLayerId(GetTestLayer()));
-
-  ASSERT_FALSE(error) << error.get();
-
-  std::promise<StreamLayerClient::FlushResponse> response_promise;
-  bool call_is_async = true;
-  auto cancel_token =
-      client_->Flush([&](StreamLayerClient::FlushResponse response) {
-        call_is_async = false;
-        response_promise.set_value(response);
-      });
-
-  EXPECT_TRUE(call_is_async);
-  auto response_future = response_promise.get_future();
-  auto status = response_future.wait_for(std::chrono::seconds(30));
-  if (status != std::future_status::ready) {
-    cancel_token.cancel();
-  }
-  auto response = response_future.get();
-
-  ASSERT_FALSE(response.empty());
-  ASSERT_NO_FATAL_FAILURE(PublishDataSuccessAssertions(response[0]));
-}
-
-TEST_F(StreamLayerClientCacheOnlineTest, FlushDataMultipleAsync) {
-  ASSERT_NO_FATAL_FAILURE(QueueMultipleEvents(5));
-
-  std::promise<StreamLayerClient::FlushResponse> response_promise;
-  bool call_is_async = true;
-  auto cancel_token =
-      client_->Flush([&](StreamLayerClient::FlushResponse response) {
-        call_is_async = false;
-        response_promise.set_value(response);
-      });
-
-  EXPECT_TRUE(call_is_async);
-  auto response_future = response_promise.get_future();
-  auto status = response_future.wait_for(std::chrono::seconds(30));
-  if (status != std::future_status::ready) {
-    cancel_token.cancel();
-  }
-  auto response = response_future.get();
-
-  ASSERT_EQ(5, response.size());
-  for (auto& single_response : response) {
-    ASSERT_NO_FATAL_FAILURE(PublishDataSuccessAssertions(single_response));
-  }
-}
-
-TEST_F(StreamLayerClientCacheOnlineTest, FlushDataCancel) {
-  auto error = client_->Queue(
-      PublishDataRequest().WithData(data_).WithLayerId(GetTestLayer()));
-
-  ASSERT_FALSE(error) << error.get();
-
-  auto cancel_future = client_->Flush();
-
-  std::thread([cancel_future]() {
-    std::this_thread::sleep_for(std::chrono::milliseconds(200));
-    cancel_future.GetCancellationToken().cancel();
-  }).detach();
-
-  auto response = cancel_future.GetFuture().get();
-
-  ASSERT_EQ(1, response.size());
-  if (response[0].IsSuccessful()) {
-    SUCCEED();
-    return;
-  }
-
-  ASSERT_NO_FATAL_FAILURE(PublishFailureAssertions(response[0]));
-}
-
-TEST_F(StreamLayerClientCacheOnlineTest, FlushListenerMetrics) {
-  disk_cache_->Close();
-  flush_settings_.auto_flush_num_events = 3;
-  client_ = CreateStreamLayerClient();
-
-  ASSERT_NO_FATAL_FAILURE(QueueMultipleEvents(3));
-
-  auto default_listener = StreamLayerClient::DefaultListener();
-
-  client_->Enable(default_listener);
-
-  for (int i = 0; default_listener->GetNumFlushEvents() < 1; i++) {
-    std::this_thread::sleep_for(std::chrono::milliseconds(50));
-    if (i > 200) {
-      FAIL() << "Timeout waiting for Flush Event Listener Results";
-    }
-  }
-
-  EXPECT_EQ(1, default_listener->GetNumFlushEvents());
-  EXPECT_EQ(1, default_listener->GetNumFlushEventsAttempted());
-  EXPECT_EQ(0, default_listener->GetNumFlushEventsFailed());
-  EXPECT_EQ(3, default_listener->GetNumFlushedRequests());
-  EXPECT_EQ(0, default_listener->GetNumFlushedRequestsFailed());
-}
-
-TEST_F(StreamLayerClientCacheOnlineTest,
-       FlushListenerMetricsSetListenerBeforeQueuing) {
-  disk_cache_->Close();
-  flush_settings_.auto_flush_num_events = 3;
-  client_ = CreateStreamLayerClient();
-
-  auto default_listener = StreamLayerClient::DefaultListener();
-  client_->Enable(default_listener);
-
-  ASSERT_NO_FATAL_FAILURE(QueueMultipleEvents(3));
-
-  for (int i = 0; default_listener->GetNumFlushEvents() < 1; i++) {
-    std::this_thread::sleep_for(std::chrono::milliseconds(50));
-    if (i > 200) {
-      FAIL() << "Timeout waiting for Flush Event Listener Results";
-    }
-  }
-
-  EXPECT_EQ(1, default_listener->GetNumFlushEvents());
-  EXPECT_EQ(1, default_listener->GetNumFlushEventsAttempted());
-  EXPECT_EQ(0, default_listener->GetNumFlushEventsFailed());
-  EXPECT_EQ(3, default_listener->GetNumFlushedRequests());
-  EXPECT_EQ(0, default_listener->GetNumFlushedRequestsFailed());
-}
-
-TEST_F(StreamLayerClientCacheOnlineTest, FlushListenerDisable) {
-  disk_cache_->Close();
-  flush_settings_.auto_flush_num_events = 3;
-  client_ = CreateStreamLayerClient();
-
-  auto default_listener = StreamLayerClient::DefaultListener();
-  client_->Enable(default_listener);
-
-  ASSERT_NO_FATAL_FAILURE(QueueMultipleEvents(3));
-
-  std::this_thread::sleep_for(std::chrono::milliseconds(100));
-
-  auto disable_future = client_->Disable();
-  auto status = disable_future.wait_for(std::chrono::seconds(5));
-  if (status != std::future_status::ready) {
-    FAIL() << "Timeout waiting for auto flushing to be disabled";
-  }
-  disable_future.get();
-
-  EXPECT_EQ(1, default_listener->GetNumFlushEvents());
-  EXPECT_EQ(1, default_listener->GetNumFlushEventsAttempted());
-  EXPECT_EQ(1, default_listener->GetNumFlushEventsFailed());
-}
-
-TEST_F(StreamLayerClientCacheOnlineTest,
-       FlushListenerMetricsMultipleFlushEventsInSeries) {
-  disk_cache_->Close();
-  flush_settings_.auto_flush_num_events = 2;
-  client_ = CreateStreamLayerClient();
-
-  auto default_listener = StreamLayerClient::DefaultListener();
-  client_->Enable(default_listener);
-
-  ASSERT_NO_FATAL_FAILURE(QueueMultipleEvents(2));
-
-  for (int i = 0, j = 1;; i++) {
-    std::this_thread::sleep_for(std::chrono::milliseconds(50));
-    if (default_listener->GetNumFlushEvents() == j) {
-      if (j == 3) {
-        break;
-      }
-      ASSERT_NO_FATAL_FAILURE(QueueMultipleEvents(2));
-      j++;
-    }
-    if (i > 400) {
-      FAIL() << "Timeout waiting for Flush Event Listener Results";
-    }
-  }
-
-  EXPECT_EQ(3, default_listener->GetNumFlushEvents());
-  EXPECT_EQ(3, default_listener->GetNumFlushEventsAttempted());
-  EXPECT_EQ(0, default_listener->GetNumFlushEventsFailed());
-  EXPECT_EQ(6, default_listener->GetNumFlushedRequests());
-  EXPECT_EQ(0, default_listener->GetNumFlushedRequestsFailed());
-}
-
-TEST_F(StreamLayerClientCacheOnlineTest,
-       FlushListenerMetricsMultipleFlushEventsInParallel) {
-  disk_cache_->Close();
-  flush_settings_.auto_flush_num_events = 2;
-  flush_settings_.events_per_single_flush =
-      flush_settings_.auto_flush_num_events;
-  client_ = CreateStreamLayerClient();
-
-  auto default_listener = StreamLayerClient::DefaultListener();
-  client_->Enable(default_listener);
-
-  ASSERT_NO_FATAL_FAILURE(QueueMultipleEvents(6));
-
-  for (int i = 0; default_listener->GetNumFlushedRequests() < 6; i++) {
-    std::this_thread::sleep_for(std::chrono::milliseconds(50));
-    if (i > 500) {
-      FAIL() << "Timeout waiting for Flush Event Listener Results";
-    }
-  }
-
-  EXPECT_LE(3, default_listener->GetNumFlushEvents());
-  EXPECT_LE(3, default_listener->GetNumFlushEventsAttempted());
-  // TODO: Investigate why there are more triggeres in auto flushing then
-  // requests. Seems AutoFlushController is trying to flush to often.
-  // EXPECT_EQ(0, default_listener->GetNumFlushEventsFailed());
-  EXPECT_EQ(6, default_listener->GetNumFlushedRequests());
-  EXPECT_EQ(0, default_listener->GetNumFlushedRequestsFailed());
-}
-
-TEST_F(StreamLayerClientCacheOnlineTest,
-       FlushListenerMetricsMultipleFlushEventsInParallelStaggeredQueue) {
-  disk_cache_->Close();
-  flush_settings_.auto_flush_num_events = 2;
-  flush_settings_.events_per_single_flush =
-      flush_settings_.auto_flush_num_events;
-  client_ = CreateStreamLayerClient();
-
-  auto default_listener = StreamLayerClient::DefaultListener();
-  client_->Enable(default_listener);
-
-  ASSERT_NO_FATAL_FAILURE(QueueMultipleEvents(4));
-  std::this_thread::sleep_for(std::chrono::milliseconds(50));
-  ASSERT_NO_FATAL_FAILURE(QueueMultipleEvents(2));
-  std::this_thread::sleep_for(std::chrono::milliseconds(100));
-  ASSERT_NO_FATAL_FAILURE(QueueMultipleEvents(4));
-
-  for (int i = 0; default_listener->GetNumFlushedRequests() < 10; i++) {
-    std::this_thread::sleep_for(std::chrono::milliseconds(50));
-    if (i > 600) {
-      FAIL() << "Timeout waiting for Flush Event Listener Results";
-    }
-  }
-
-  EXPECT_LE(3, default_listener->GetNumFlushEvents());
-  EXPECT_LE(3, default_listener->GetNumFlushEventsAttempted());
-  // TODO: Investigate why there are more triggeres in auto flushing then
-  // requests. Seems AutoFlushController is trying to flush to often.
-  // EXPECT_EQ(0, default_listener->GetNumFlushEventsFailed());
-  EXPECT_EQ(10, default_listener->GetNumFlushedRequests());
-  EXPECT_EQ(0, default_listener->GetNumFlushedRequestsFailed());
-}
-
-TEST_F(StreamLayerClientCacheOnlineTest, FlushListenerNotifications) {
-  disk_cache_->Close();
-  flush_settings_.auto_flush_num_events = 3;
-  client_ = CreateStreamLayerClient();
-
-  ASSERT_NO_FATAL_FAILURE(QueueMultipleEvents(3));
-
-  class NotificationListener : public DefaultFlushEventListener<
-                                   const StreamLayerClient::FlushResponse&> {
-   public:
-    void NotifyFlushEventStarted() override { events_started_++; }
-
-    void NotifyFlushEventResults(
-        const StreamLayerClient::FlushResponse& results) override {
-      std::lock_guard<std::mutex> lock(results_mutex_);
-      results_ = std::move(results);
-    }
-
-    const StreamLayerClient::FlushResponse& GetResults() {
-      std::lock_guard<std::mutex> lock(results_mutex_);
-      return results_;
-    }
-
-    int events_started_ = 0;
-
-   private:
-    StreamLayerClient::FlushResponse results_;
-    std::mutex results_mutex_;
-  };
-
-  auto notification_listener = std::make_shared<NotificationListener>();
-  client_->Enable(notification_listener);
-
-  for (int i = 0; notification_listener->GetResults().size() < 3; i++) {
-    std::this_thread::sleep_for(std::chrono::milliseconds(50));
-    if (i > 200) {
-      FAIL() << "Timeout waiting for Flush Event Listener Results";
-    }
-  }
-
-  EXPECT_EQ(1, notification_listener->events_started_);
-  for (auto result : notification_listener->GetResults()) {
-    ASSERT_NO_FATAL_FAILURE(PublishDataSuccessAssertions(result));
-  }
-}
-
-TEST_F(StreamLayerClientCacheOnlineTest, FlushSettingsTimeSinceOldRequest) {
-  disk_cache_->Close();
-  flush_settings_.auto_flush_old_events_force_flush_interval = 10;
-  client_ = CreateStreamLayerClient();
-
-  ASSERT_NO_FATAL_FAILURE(QueueMultipleEvents(2));
-
-  auto default_listener = StreamLayerClient::DefaultListener();
-  client_->Enable(default_listener);
-
-  for (int i = 0; default_listener->GetNumFlushEvents() < 1; i++) {
-    std::this_thread::sleep_for(std::chrono::milliseconds(50));
-    if (i > 400) {
-      FAIL() << "Timeout waiting for Flush Event Listener Results";
-    }
-  }
-
-  EXPECT_EQ(1, default_listener->GetNumFlushEvents());
-  EXPECT_EQ(1, default_listener->GetNumFlushEventsAttempted());
-  EXPECT_EQ(0, default_listener->GetNumFlushEventsFailed());
-  EXPECT_EQ(2, default_listener->GetNumFlushedRequests());
-  EXPECT_EQ(0, default_listener->GetNumFlushedRequestsFailed());
-}
-
-TEST_F(StreamLayerClientCacheOnlineTest,
-       FlushSettingsTimeSinceOldRequestQueueAfterEnable) {
-  disk_cache_->Close();
-  flush_settings_.auto_flush_old_events_force_flush_interval = 10;
-  client_ = CreateStreamLayerClient();
-
-  auto default_listener = StreamLayerClient::DefaultListener();
-  client_->Enable(default_listener);
-
-  ASSERT_NO_FATAL_FAILURE(QueueMultipleEvents(2));
-
-  for (int i = 0; default_listener->GetNumFlushEvents() < 1; i++) {
-    std::this_thread::sleep_for(std::chrono::milliseconds(50));
-    if (i > 400) {
-      FAIL() << "Timeout waiting for Flush Event Listener Results";
-    }
-  }
-
-  EXPECT_EQ(1, default_listener->GetNumFlushEvents());
-  EXPECT_EQ(1, default_listener->GetNumFlushEventsAttempted());
-  EXPECT_EQ(0, default_listener->GetNumFlushEventsFailed());
-  EXPECT_EQ(2, default_listener->GetNumFlushedRequests());
-  EXPECT_EQ(0, default_listener->GetNumFlushedRequestsFailed());
-}
-
-TEST_F(StreamLayerClientCacheOnlineTest,
-       FlushSettingsTimeSinceOldRequestDisable) {
-  disk_cache_->Close();
-  flush_settings_.auto_flush_old_events_force_flush_interval = 2;
-  client_ = CreateStreamLayerClient();
-
-  ASSERT_NO_FATAL_FAILURE(QueueMultipleEvents(2));
-
-  auto default_listener = StreamLayerClient::DefaultListener();
-  client_->Enable(default_listener);
-
-  std::this_thread::sleep_for(std::chrono::milliseconds(2100));
-
-  auto disable_future = client_->Disable();
-  auto status = disable_future.wait_for(std::chrono::seconds(5));
-  if (status != std::future_status::ready) {
-    FAIL() << "Timeout waiting for auto flushing to be disabled";
-  }
-  disable_future.get();
-
-  EXPECT_EQ(1, default_listener->GetNumFlushEvents());
-  EXPECT_EQ(1, default_listener->GetNumFlushEventsAttempted());
-  EXPECT_EQ(1, default_listener->GetNumFlushEventsFailed());
-}
-
-TEST_F(StreamLayerClientCacheOnlineTest, FlushSettingsAutoFlushInterval) {
-  disk_cache_->Close();
-  flush_settings_.auto_flush_interval = 10;
-  client_ = CreateStreamLayerClient();
-
-  ASSERT_NO_FATAL_FAILURE(QueueMultipleEvents(2));
-
-  auto default_listener = StreamLayerClient::DefaultListener();
-  client_->Enable(default_listener);
-
-  for (int i = 0; default_listener->GetNumFlushEvents() < 1; i++) {
-    std::this_thread::sleep_for(std::chrono::milliseconds(50));
-    if (i > 400) {
-      FAIL() << "Timeout waiting for Flush Event Listener Results";
-    }
-  }
-
-  EXPECT_EQ(1, default_listener->GetNumFlushEvents());
-  EXPECT_EQ(1, default_listener->GetNumFlushEventsAttempted());
-  EXPECT_EQ(0, default_listener->GetNumFlushEventsFailed());
-  EXPECT_EQ(2, default_listener->GetNumFlushedRequests());
-  EXPECT_EQ(0, default_listener->GetNumFlushedRequestsFailed());
-}
-
-TEST_F(StreamLayerClientCacheOnlineTest,
-       FlushSettingsAutoFlushIntervalDisable) {
-  disk_cache_->Close();
-  flush_settings_.auto_flush_interval = 2;
-  client_ = CreateStreamLayerClient();
-
-  ASSERT_NO_FATAL_FAILURE(QueueMultipleEvents(2));
-
-  auto default_listener = StreamLayerClient::DefaultListener();
-  client_->Enable(default_listener);
-
-  std::this_thread::sleep_for(std::chrono::milliseconds(2100));
-
-  auto disable_future = client_->Disable();
-  auto status = disable_future.wait_for(std::chrono::seconds(5));
-  if (status != std::future_status::ready) {
-    FAIL() << "Timeout waiting for auto flushing to be disabled";
-  }
-  disable_future.get();
-
-  EXPECT_EQ(1, default_listener->GetNumFlushEvents());
-  EXPECT_EQ(1, default_listener->GetNumFlushEventsAttempted());
-  EXPECT_EQ(1, default_listener->GetNumFlushEventsFailed());
-}
-
 }  // namespace
